+++ src/Symfony/Component/Validator/Context/ExecutionContext.php
@@ -13,6 +13,7 @@ namespace Symfony\Component\Validator\Context;
 
 use Symfony\Component\Translation\TranslatorInterface;
 use Symfony\Component\Validator\ClassBasedInterface;
+use Symfony\Component\Validator\Constraint;
 use Symfony\Component\Validator\ConstraintViolation;
 use Symfony\Component\Validator\ConstraintViolationList;
 use Symfony\Component\Validator\Exception\BadMethodCallException;
@@ -100,6 +101,13 @@ class ExecutionContext implements ExecutionContextInterface
     private $group;
 
     /**
+     * The currently validated constraint.
+     *
+     * @var Constraint|null
+     */
+    private $constraint;
+
+    /**
      * Stores which objects have been validated in which group.
      *
      * @var array
@@ -165,6 +173,14 @@ class ExecutionContext implements ExecutionContextInterface
     /**
      * {@inheritdoc}
      */
+    public function setConstraint(Constraint $constraint)
+    {
+        $this->constraint = $constraint;
+    }
+
+    /**
+     * {@inheritdoc}
+     */
     public function addViolation($message, array $parameters = array(), $invalidValue = null, $plural = null, $code = null)
     {
         // The parameters $invalidValue and following are ignored by the new
@@ -186,7 +202,8 @@ class ExecutionContext implements ExecutionContextInterface
             $this->propertyPath,
             $this->value,
             null,
-            null
+            null,
+            $this->constraint
         ));
     }
 
@@ -197,6 +214,7 @@ class ExecutionContext implements ExecutionContextInterface
     {
         return new ConstraintViolationBuilder(
             $this->violations,
+            $this->constraint,
             $message,
             $parameters,
             $this->root,